Improved zlib wrapper
Moved implementation to another file
Generally simplified implementation
