+++ src/hexicord/gateway_client.cpp
@@ -23,8 +23,7 @@
 #include <hexicord/config.hpp>
 #include <hexicord/internal/utils.hpp>
 #ifdef HEXICORD_ZLIB
-#include <zlib.h>
-#define ZLIB_COMPLETE_CHUNK 16384
+#include <hexicord/internal/zlib.hpp>
 #endif
 
 #if defined(HEXICORD_DEBUG_LOG)
@@ -49,60 +48,6 @@
     #define OS_STR "unknown"
 #endif
 
-#ifdef HEXICORD_ZLIB
-class ZLibDecompressor {
-public:
-    ZLibDecompressor() {
-        int retval;
-        strm_.zalloc = Z_NULL;
-        strm_.zfree = Z_NULL;
-        strm_.opaque = Z_NULL;
-        strm_.avail_in = 0;
-        strm_.next_in = Z_NULL;
-        retval = inflateInit2(&strm_, 15);
-        if (retval != Z_OK) {
-            DEBUG_MSG("Something gone wrong during zlib initialization!")
-        }
-    }
-
-    ~ZLibDecompressor() {
-        inflateEnd(&strm_);
-    }
-
-    uint8_t* decompress(const std::vector<uint8_t>& input) {
-        int retval;
-        std::vector<uint8_t> result(ZLIB_COMPLETE_CHUNK);
-        for (uint32_t i = 0; i < input.size(); i += ZLIB_COMPLETE_CHUNK) {
-            int howManyLeft = input.size() - i;
-            int howManyWanted = (howManyLeft > ZLIB_COMPLETE_CHUNK) ? ZLIB_COMPLETE_CHUNK : howManyLeft;
-            memcpy(in_, input.data() + i, howManyWanted);
-            strm_.avail_in = howManyWanted;
-            strm_.next_in = (Bytef *) in_;
-            if (strm_.avail_in == 0) {
-                break;
-            }
-            do {
-                memset(&out_, 0, ZLIB_COMPLETE_CHUNK);
-                strm_.avail_out = ZLIB_COMPLETE_CHUNK;
-                strm_.next_out = (Bytef *) out_;
-                retval = inflate(&strm_, Z_NO_FLUSH);
-                if (retval == Z_STREAM_ERROR) {
-                    DEBUG_MSG("zlib stream error!")
-                }
-                for (int j = 0; j < ZLIB_COMPLETE_CHUNK; j++) {
-                    result.push_back(out_[j]);
-                }
-            } while (strm_.avail_out == 0);
-        }
-        return result.data();
-    }
-private:
-    char in_[ZLIB_COMPLETE_CHUNK];
-    char out_[ZLIB_COMPLETE_CHUNK];
-    z_stream strm_;
-};
-#endif
-
 namespace Hexicord {
 
 GatewayClient::GatewayClient(boost::asio::io_service& ioService, const std::string& token)
@@ -117,11 +62,7 @@ nlohmann::json GatewayClient::parseGatewayMessage(const std::vector<uint8_t>& ms
     if (msg[0] == '{') {
         return nlohmann::json::parse(msg);
     }
-    uint8_t* carr;
-    {
-        ZLibDecompressor decompressor = ZLibDecompressor();
-        carr = decompressor.decompress(msg);
-    }
+    uint8_t* carr = zlib::decompress(msg);
     return nlohmann::json::parse(carr);
 #else
     return nlohmann::json::parse(msg);
@@ -150,7 +91,7 @@ void GatewayClient::connect(const std::string& gatewayUrl, int shardId, int shar
             { "device", "hexicord" }
         }},
 #ifdef HEXICORD_ZLIB
-        { "compress", true}
+        { "compress", true },
 #else
         { "compress", false },
 #endif
